Units Conversion Functions
Convert input parameters and concentration units to base units. Convert output units back to the input ones if nothing else is specified. If flow rate and linear velocity are specified it takes only the linear velocity and converts it to the base units. Otherwise converts flow rate to velocity in base units.